adding a remove header function
This might be makes no sense at all because it is 2.x version of fof but this is not the first time that try to remove a header field from a browse form and failed.

So I copied the removeField function from JForm and changed so that it ow removes headers.