~ Optimized the way the default media query is determined
We now use a MediaVersion service instead of a simple string.

The MediaVersion object can be converted to string automatically
to preserve backwards compatibility.

For the same b/c reasons we have modified the Container to proxy
setting the mediaVersion in the container to calling the
setMediaVersion() method on the MediaVersion object.

Moreover, if you've set the mediaVersion key in the container's
configuration it is automatically proxied to the MediaVersion
object. Same goes with fof.xml.

From the developer's perspective nothing has changed. Internally,
it's a whole new level which saves a few database queries when
you are overriding the mediaVersion in the Dispatcher.